AudioLevels: Better analysis logic for peaks
This is perhaps still not quite perfect, but it is much better than the
old logic (no more int/float juggling, just a conversion to dB for
display purposes, and also just leveraging the hard work done to
optimize the min/max discovery logic juce has already done for us).

+++ src/AudioLevelsChannel.h
@@ -39,8 +39,8 @@ public:
     jack_client_t *jackClient{nullptr};
     float peakAHoldSignal{0};
     float peakBHoldSignal{0};
-    int peakA{0};
-    int peakB{0};
+    float peakA{0};
+    float peakB{0};
     bool enabled{false};
     QString clientName;
     quint64 firstRecordingFrame{0};
